Update README.md
DynaPDF 4.0.94.265 (x86/x64)

Anyone who needs to edit or create PDF files on client / server systems, requires a PDF library that can deal with current PDF files in the future. The PDF format was greatly extended in recent years and the development does never end. A PDF library must keep pace with this development but it is almost impossible to implement all new features immediately after the publication of a new version of PDF. Therefore, a PDF library should ensure that new features do not lead to errors or even data loss.

[Download DynaPDF 4.0.94.265 (x86/x64)](https://developer.team/dotnet/35128-dynapdf-4094265-x86x64.html)